feat: added the gasPrice
