+++ src/components/TransactionBlock/AssetSwapTransactionBlock.tsx
@@ -21,6 +21,7 @@ import { swapServiceIdToDetails } from '../../utils/swap';
 import Text from '../Text/Text';
 import { IAssetSwapTransactionBlock, IMultiCallData } from '../../types/transactionBlock';
 import useAssetPriceUsd from '../../hooks/useAssetPriceUsd';
+import { OfferRoute } from '../OfferRoute/OfferRoute';
 
 export interface ISwapAssetTransactionBlockValues {
   chain?: Chain;
@@ -33,30 +34,12 @@ export interface ISwapAssetTransactionBlockValues {
   accountType?: string;
 }
 
-type RouteOptionProps = {
-  option: SelectOption;
-  availableOffers: ExchangeOffer[] | null;
-  availableToAssets: TokenListToken[] | null;
-  selectedToAsset: TokenListToken | null;
-  targetAssetPriceUsd: number | null;
-  selectedFromAsset: IAssetWithBalance | null;
-  selectedNetwork: Chain | null;
-  selectedAccountType: string;
-};
-
 const Title = styled.h3`
   margin: 0 0 18px;
   padding: 0;
   font-size: 16px;
   color: ${({ theme }) => theme.color.text.cardTitle};
   font-family: 'PTRootUIWebBold', sans-serif;
-`;
-
-const OfferDetails = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  font-family: 'PTRootUIWebMedium', sans-serif;
 `;
 
 const mapOfferToOption = (offer: ExchangeOffer) => {
@@ -73,84 +56,6 @@ const mapAssetToOption = (asset: TokenListToken) => ({
   value: asset.address,
   iconUrl: asset.logoURI,
 });
-
-const RenderOption = (props: RouteOptionProps) => {
-  const {
-    option,
-    availableOffers,
-    availableToAssets,
-    selectedToAsset,
-    targetAssetPriceUsd,
-    selectedNetwork,
-    selectedFromAsset,
-    selectedAccountType,
-  } = props;
-  const [gasPrice, setGasPrice] = useState('-');
-  const [isEstimating, setIsEstimating] = useState(true);
-  const { getSdkForChainId } = useEtherspot();
-
-  const getGasSwapUsdValue = async (offer: ExchangeOffer) => {
-    const sdkByChain = getSdkForChainId(selectedNetwork?.chainId ?? 1);
-    setIsEstimating(true);
-
-    if (sdkByChain && selectedFromAsset && selectedAccountType === AccountTypes.Contract) {
-      await sdkByChain.computeContractAccount();
-
-      await sdkByChain.clearGatewayBatch();
-
-      for (let i = 0; i < offer.transactions.length; i++) {
-        await sdkByChain.batchExecuteAccountTransaction(offer.transactions[i]);
-      }
-
-      const feeTokens = await sdkByChain.getGatewaySupportedTokens();
-
-      try {
-        const estimation = await sdkByChain.estimateGatewayBatch({ feeToken: feeTokens[0].address }); // pay gas using USDC
-        setIsEstimating(false);
-        return `${ethers.utils.formatUnits(estimation.estimation.feeAmount)}`;
-      } catch (error) {
-        //
-      }
-    }
-
-    setIsEstimating(false);
-    return '-';
-  };
-
-  const availableOffer = availableOffers?.find((offer) => offer.provider === option.value);
-  const toAsset = availableToAssets?.find((availableAsset) =>
-    addressesEqual(availableAsset.address, selectedToAsset?.address)
-  );
-
-  const valueToReceiveRaw = availableOffer
-    ? ethers.utils.formatUnits(availableOffer.receiveAmount, toAsset?.decimals)
-    : undefined;
-
-  const valueToReceive = valueToReceiveRaw && formatAmountDisplay(valueToReceiveRaw);
-
-  useEffect(() => {
-    if (availableOffer) {
-      getGasSwapUsdValue(availableOffer).then((res) => setGasPrice(formatAmountDisplay(res, '$')));
-    }
-  }, [availableOffer]);
-
-  return (
-    <OfferDetails>
-      <RoundedImage title={option.title} url={option.iconUrl} size={24} />
-      <div>
-        <Text size={12} marginBottom={2} medium block>
-          {option.title} gasPrice: {isEstimating ? 'Estimating...' : gasPrice}
-        </Text>
-        {!!valueToReceive && (
-          <Text size={16} medium>
-            {valueToReceive} {toAsset?.symbol}
-            {targetAssetPriceUsd && ` · ${formatAmountDisplay(+valueToReceiveRaw * targetAssetPriceUsd, '$')}`}
-          </Text>
-        )}
-      </div>
-    </OfferDetails>
-  );
-};
 
 const AssetSwapTransactionBlock = ({
   id: transactionBlockId,
@@ -479,7 +384,7 @@ const AssetSwapTransactionBlock = ({
             setSelectedOffer(option);
           }}
           renderOptionListItemContent={(option) => (
-            <RenderOption
+            <OfferRoute
               option={option}
               availableOffers={availableOffers}
               availableToAssets={availableToAssets}
@@ -491,7 +396,7 @@ const AssetSwapTransactionBlock = ({
             />
           )}
           renderSelectedOptionContent={(option) => (
-            <RenderOption
+            <OfferRoute
               option={option}
               availableOffers={availableOffers}
               availableToAssets={availableToAssets}
